refactor : 체스판 한 칸의 정보를 담는 Cell 클래스 관련 수정
- DTO 역할을 하므로 패키지를 DTO로 변경
- Cell 클래스가 Board 정보(위치&말)를 받으면 내부에서 필요한 형태로 바꿀 수 있도록 생성자와 메서드 추가
- Cell 클래스에서 체스칸의 색깔을 구할 때 필요하기 때문에, Position 클래스에 좌표의 합을 구해주는 메서드 추가
- 도메인 Board에 꼭 필요한 기능만 두기 위해, Board가 갖고 있던 Cell 리스트를 만드는 기능을 Service 클래스로 옮김
