+++ src/main/java/chess/domain/board/Board.java
@@ -3,10 +3,8 @@ package chess.domain.board;
 import chess.domain.GameResult;
 import chess.domain.command.MoveCommand;
 import chess.domain.piece.*;
-import chess.dto.BoardDTO;
 import chess.exception.*;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -16,10 +14,6 @@ import static chess.domain.piece.Pawn.PAWN_SCORE;
 import static chess.util.NullValidator.validateNull;
 
 public class Board {
-    private static final char MIN_Y_POINT = '1';
-    private static final char MAX_Y_POINT = '8';
-    private static final char MIN_X_POINT = 'a';
-    private static final char MAX_X_POINT = 'h';
     public static final int ONLY_ONE_PAWN_IN_XPOINT = 1;
 
     private Map<Position, Piece> board;
@@ -191,61 +185,5 @@ public class Board {
 
     public Map<Position, Piece> get() {
         return Collections.unmodifiableMap(board);
-    }
-
-    public List<Cell> getCells() {
-        List<Cell> cells = new ArrayList<>();
-
-        for (char yPoint = MAX_Y_POINT; yPoint >= MIN_Y_POINT; yPoint--) {
-            for (char xPoint = MIN_X_POINT; xPoint <= MAX_X_POINT; xPoint++) {
-                Piece piece = this.board.get(PositionFactory.of(xPoint, yPoint));
-
-                Cell cell = new Cell();
-                cell.setTileColor(getTileColor(xPoint, yPoint));
-                cell.setPosition(String.valueOf(xPoint) + (yPoint));
-                cell.setPieceColor(piece.getPieceColorName());
-                cell.setPiece(piece.getName());
-                cells.add(cell);
-            }
-        }
-
-        return cells;
-    }
-
-    private String getTileColor(char xPoint, char yPoint) {
-        int x = xPoint - MIN_X_POINT;
-        int y = yPoint - MIN_Y_POINT;
-
-        if ((x + y) % 2 == 0) {
-            return "BLACK";
-        }
-        if ((x + y) % 2 != 0) {
-            return "WHITE";
-        }
-        return "NONE";
-    }
-
-    public BoardDTO createBoardDTO() {
-        BoardDTO boardDTO = new BoardDTO();
-
-        boardDTO.setLines(getBoardLines());
-        boardDTO.setCurrentTeam(this.team.getName());
-
-        return boardDTO;
-    }
-
-    private List<String> getBoardLines() {
-        List<String> lines = new ArrayList<>();
-
-        for (char yPoint = MIN_Y_POINT; yPoint <= MAX_Y_POINT; yPoint++) {
-            StringBuilder stringBuilder = new StringBuilder();
-            for (char xPoint = MIN_X_POINT; xPoint <= MAX_X_POINT; xPoint++) {
-                Piece piece = this.board.get(PositionFactory.of(xPoint, yPoint));
-                stringBuilder.append(piece.getName());
-            }
-            lines.add(stringBuilder.toString());
-        }
-
-        return lines;
     }
 }
