--- src/main/java/chess/domain/board/Cell.java
@@ -1,40 +1,0 @@
-package chess.domain.board;
-
-public class Cell {
-    private String tileColor;
-    private String position;
-    private String pieceColor;
-    private String piece;
-
-    public String getTileColor() {
-        return tileColor;
-    }
-
-    public void setTileColor(String tileColor) {
-        this.tileColor = tileColor;
-    }
-
-    public String getPosition() {
-        return position;
-    }
-
-    public void setPosition(String position) {
-        this.position = position;
-    }
-
-    public String getPieceColor() {
-        return pieceColor;
-    }
-
-    public void setPieceColor(String pieceColor) {
-        this.pieceColor = pieceColor;
-    }
-
-    public String getPiece() {
-        return piece;
-    }
-
-    public void setPiece(String piece) {
-        this.piece = piece;
-    }
-}
