+++ src/main/java/chess/service/ChessGameService.java
@@ -4,14 +4,24 @@ import chess.dao.BoardDAO;
 import chess.dao.TurnDAO;
 import chess.domain.GameResult;
 import chess.domain.board.Board;
+import chess.domain.board.Position;
+import chess.domain.board.PositionFactory;
 import chess.domain.command.MoveCommand;
+import chess.domain.piece.Piece;
 import chess.dto.BoardDTO;
 import chess.dto.Cell;
 import chess.dto.TurnDTO;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 public class ChessGameService {
+    private static final char MIN_Y_POINT = '1';
+    private static final char MAX_Y_POINT = '8';
+    private static final char MIN_X_POINT = 'a';
+    private static final char MAX_X_POINT = 'h';
+
     private BoardDAO boardDAO;
     private TurnDAO turnDAO;
     private Board board;
@@ -38,7 +48,20 @@ public class ChessGameService {
     }
 
     public List<Cell> getCells() {
-        return this.board.getCells();
+        List<Cell> cells = new ArrayList<>();
+
+        Map<Position, Piece> boardData = this.board.get();
+
+        for (char yPoint = MAX_Y_POINT; yPoint >= MIN_Y_POINT; yPoint--) {
+            for (char xPoint = MIN_X_POINT; xPoint <= MAX_X_POINT; xPoint++) {
+                Position position = PositionFactory.of(xPoint, yPoint);
+                Piece piece = boardData.get(position);
+
+                cells.add(Cell.from(position, piece));
+            }
+        }
+
+        return cells;
     }
 
     public String getCurrentTeam() {
